Clarified branch() method
The text in the README could be read as saying that calling g.branch('new_branch') creates a branch in the repository. However, it only creates a Git::Branch object. To create a branch in the repository, a user needs to call g.branch('new_branch').create or g.branch('new_branch').checkout.