Assume strings are utf8 by default and fallback on encoding if necessary
Before this patch, we always assumed the string was ascii-8bit which
does not work correctly with real utf8 strings (french accented chars or
chineese chars).

This patch tries to use the string as utf8 and fallback if necessary to
the the ascii-8bit assumption if necessary.

This should help with #295 and might replace the need for #301.

Change-Id: Idac63fa10e5aefafa1eb99a6be4138cac5f90ea0