Fix hash keys in #describe
Currently, the symbolized key is checked for presence and then the value for the non-symbolized one is used. This works for Hashes with indifferent access (like in rails etc.) but fails for a simple example like

```ruby
Git.open(".").describe("HEAD", abbrev: 0)

=> Git::GitExecuteError (git '--git-dir=/folder/to/repo/.git' '--work-tree=/folder/to/repo' describe '--abbrev=' 'HEAD'  2>&1:error: option `abbrev' expects a numerical value)
```

Changed all references to the symbolized version (since the method signature suggests `@param [{Symbol=>Object}]`).

Signed-off-by: Jonas Mueller <jonas-mueller@mailbox.org>