Disable GPG Signing in Test Config (#467)
If a user has GPG signing set to true in their global .gitconfig file,
the test suite will fail. This is because the test suit now includes a
check of a commit message (the `test_commit_with_no_verify` test in
`tests/units/test_lib.rb`) but did not update the test config file. This
commit updates the config file to set the `gpgsign` option to false.

Signed-off-by: Michael Camilleri <mike@inqk.net>