Update index before fetching modified files (#409)
According to the [git-diff-index docs][1], git commands like diff-index
and diff-files typically don't look at the actual contents of files in
the working tree. This means that these commands can be easily fooled by
operations that don't change the content of a file such as deleting and
recreating it, or even running `touch` on a file.

The docs then say that one may update the index to make it be in sync
with the actual contents of the working directory. The recommended
command would seem to be `git update-index --refresh`, but in my
testing, this caused "needs update" errors in some of the tests. So I've
resorted to using `git status`, which is probably a slightly more
heavyweight command, but does pass the tests.

[1]: https://git-scm.com/docs/git-diff-index

Signed-off-by: Gabriel Gilder <gabriel@gabrielgilder.com>