Add --all flag to fetch method
The fix for CVE-2022-25648 broke a workflow for us by removing the ability
to fetch all remotes.

This adds --all to the list of available flags for the #fetch method. It does
not add :a as an alternative flag, as git fetch -a and git fetch --all are
not the same thing.

Signed-off-by: Matthew Riedel <matthew.riedel@fluxfederation.com>