Remove #process_commmit_data indent parameter
This was introduced in [1] to handle variance in indentation when
parsing commit messages; at the time the #process_commit_data method had
two callers, one which used it to process log lines, and the other which
used it to process the output of `git cat-file commit <SHA>`. The former
sees a 4-space indent, the latter, zero.

Since [2], however, the log processing has been handled by
the #process_commit_log_data method, so #process_commit_data exclusively
handles un-indented inputs, and we can remove the indent parameter.

[1] https://github.com/ruby-git/ruby-git/commit/05117d4ebc1ba903083802f703739c55207fdfd5
[2] https://github.com/ruby-git/ruby-git/commit/97e1fff9edf0ee8cc7b49e49bd7faa4b1de16785

Signed-off-by: Simon Coffey <simon.coffey@futurelearn.com>