Use dynamically-created repo for signed commits test
This replaces the test repo introduced in #610 with one created on the
fly in the test.

I've switched from GPG to SSH signing to minimise the likelihood that
we'll need extra dependencies on CI or other contributors' machines.

I've also removed some of the commit metadata assertions that didn't
feel particularly necessary, and this allowed me to reduce the setup
steps for clarity.

Signed-off-by: Simon Coffey <simon.coffey@futurelearn.com>