Supply all of the _specific_ color options too
Previously, we were supplying `color.ui=false`, but if the local
gitconfig specified any of the more specific options (like
`color.diff`), those would take precedence. This updates our
command-runner to always supply all of the specific color options as
false as well, so that we definitely get a color-free output suitable
for parsing.