Memoize all of the significant calls in Git::Status
When the status has many entries, there were substantial inefficiencies
in this class - calling predicates like `changed?(filename)` would
iterate the status, constructing a transient `changed` subhash, then
test that subhash to see if the file in question was in it (for
example).

After this, it will _keep_ those sub-hashes for reuse on the Status
instance, as well as downcased versions if they happen to get requested
(by case-insensitive calls).