fix: un-deprecate Git::Diff methods
These methods were deprecated with the same thinking that
the Git::Log methods were deprecated. However, where
Git::Log is a query builder, Git::Diff is (mostly) not... it is a
facade over diff, diff_stats, and diff_path_stats.

A problem remains with the Git::Diff class in that if you call
the #path method after retrieving results, the results are not
updated which may cause unexpected results.

I'll consider what changes should be made to the Git::Diff
class at a later date.